service to calculate account info, -account balance -income last month & this year -expense last month & this year

+++ database/migrations/2022_07_04_083404_create_accounts_table.php
@@ -24,8 +24,8 @@ return new class extends Migration
             $table->foreignIdFor(AccountType::class)->constrained();
             $table->foreignIdFor(Family::class)->nullable()->constrained();
             $table->unsignedInteger('position');
-            $table->integer('start_balance');
-            $table->integer('alert');
+            $table->float('start_balance');
+            $table->float('alert');
             $table->timestamps();
             $table->softDeletes();
         });
